pyproject.toml
[tool.poetry]
name = "multimediasorter"
version = "1.0.8"
description = ""
authors = []
license = "GPL-3.0"
readme = "README.md"
packages = [{ include = "mediasorter", from = "src" }]

[tool.poetry.scripts]
mediasorter = "mediasorter.cli:main"

[tool.poetry.dependencies]
python = ">=3.8"
PyYAML = "^6.0"
pydantic = "^1.10.4"
click = "^8.1.3"
aiohttp = "^3.8.3"
rich = "^13.1.0"
async-cache = "^1.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-asyncio = "^0.20.3"
mock = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
